requirements.txt
torch          
torchvision
tqdm==4.57.0
tensorboard==1.14.0
Pillow==8
opencv-python
